requirements.txt
blinker==1.8.2
click==8.1.7
distro==1.9.0
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.3
